Refactor: 하프 모달 뷰 컨트롤러 코드 추가 및 이동
