+++ Spacing/ViewController/DetailViewController.swift
@@ -7,83 +7,6 @@
 
 import UIKit
 
-class HalfModalPresentationController: UIPresentationController {
-    // 하프 모달의 뒤 배경 뷰
-    private var dimmingView: UIView!
-    
-    private var initialTouchPoint: CGPoint = CGPoint(x: 0, y: 0)
-    
-    override var frameOfPresentedViewInContainerView: CGRect {
-        guard let containerView = containerView else { return CGRect.zero }
-        
-        let height: CGFloat = 300 // 하프 모달의 높이
-        return CGRect(x: 0, y: containerView.bounds.height - height, width: containerView.bounds.width, height: height)
-    }
-    
-    // 추가: 드래그 제스처를 처리하는 메서드
-    @objc private func handlePanGesture(_ gestureRecognizer: UIPanGestureRecognizer) {
-        guard let presentedView = presentedView else { return }
-        
-        let translation = gestureRecognizer.translation(in: presentedView)
-        
-        switch gestureRecognizer.state {
-        case .began:
-            initialTouchPoint = presentedView.frame.origin
-        case .changed:
-            presentedView.frame.origin = CGPoint(
-                x: initialTouchPoint.x,
-                y: initialTouchPoint.y + translation.y
-            )
-        case .ended, .cancelled:
-            let velocity = gestureRecognizer.velocity(in: presentedView)
-            if velocity.y > 0 {
-                // 아래로 드래그: 모달 닫기
-                presentingViewController.dismiss(animated: true, completion: nil)
-            } else {
-                // 위로 드래그: 모달 유지
-                presentedView.frame.origin = initialTouchPoint
-            }
-        default:
-            break
-        }
-    }
-    
-    // 모달이 표시될 때 호출되는 메서드
-    override func presentationTransitionWillBegin() {
-        guard let containerView = containerView else { return }
-        
-        // 어둡게 처리할 배경 뷰 생성
-        dimmingView = UIView(frame: containerView.bounds)
-        dimmingView.backgroundColor = UIColor.black.withAlphaComponent(0.5) // 어둡게 처리할 배경의 투명도 조절
-        dimmingView.alpha = 0 // 초기에는 보이지 않도록 설정
-        containerView.insertSubview(dimmingView, at: 0)
-        
-        // 애니메이션을 사용하여 배경 뷰를 어둡게 표시
-        presentedViewController.transitionCoordinator?.animate(alongsideTransition: { _ in
-            self.dimmingView.alpha = 1 // 어둡게 처리할 배경의 투명도를 1로 변경하여 보이도록 함
-        }, completion: nil)
-        
-        // 추가: 드래그 제스처 추가
-        let panGestureRecognizer = UIPanGestureRecognizer(target: self, action: #selector(handlePanGesture(_:)))
-        presentedViewController.view.addGestureRecognizer(panGestureRecognizer)
-        
-        // 모달의 모서리를 둥글게 만듦
-        presentedViewController.view.layer.cornerRadius = 12
-        presentedViewController.view.clipsToBounds = true
-    }
-    
-    // 모달이 사라질 때 호출되는 메서드
-    override func dismissalTransitionWillBegin() {
-        // 애니메이션을 사용하여 배경 뷰를 다시 투명하게 만듦
-        presentedViewController.transitionCoordinator?.animate(alongsideTransition: { _ in
-            self.dimmingView.alpha = 0 // 어둡게 처리할 배경의 투명도를 0으로 변경하여 숨김
-        }, completion: { _ in
-            // 애니메이션 완료 후 배경 뷰를 제거
-            self.dimmingView.removeFromSuperview()
-        })
-    }
-}
-
 class DetailViewController: UIViewController, TimeSelectionDelegate {
     
     // MARK: - Properties
